feat(text-search-engine): :sparkles: export new component 'HighlightWithTarget' for react
