+++ .changeset/eight-days-cover.md
@@ -1,0 +1,5 @@
+---
+'text-search-engine': minor
+---
+
+export new component 'HighlightWithTarget' and 'HighlightWithRanges' for react
