+++ .vscode/settings.json
@@ -1,5 +1,5 @@
 {
-	"eslint.enable": false,
+	"eslint.enable": true,
 	"prettier.enable": false,
 	"biome.enabled": true,
 	"[javascript]": {
