+++ commitlint.config.js
@@ -12,6 +12,7 @@ const scopes = fg
 module.exports = {
 	extends: ['@commitlint/config-conventional'],
 	rules: {
+		'limit-description-length': [2, 'always', 120],
 		'scope-enum': [2, 'always', scopes],
 	},
 	prompt: {
