--- components/react/package.json
@@ -1,38 +1,0 @@
-{
-	"name": "@text-search-engine/react",
-	"version": "1.2.0",
-	"private": false,
-	"description": "React components for text search and highlighting",
-	"license": "MIT",
-	"author": "cjinhuo",
-	"type": "module",
-	"main": "dist/index.js",
-	"module": "dist/index.esm.js",
-	"types": "dist/index.d.ts",
-	"publishConfig": {},
-	"homepage": "https://cjinhuo.github.io/text-search-engine/",
-	"repository": {
-		"type": "git",
-		"url": "git+https://github.com/cjinhuo/text-search-engine.git"
-	},
-	"bugs": {
-		"url": "https://github.com/cjinhuo/text-search-engine/issues"
-	},
-	"files": ["dist"],
-	"scripts": {
-		"build": "rollup -c",
-		"prepublishOnly": "npm run build"
-	},
-	"dependencies": {
-		"react": "^18.3.1",
-		"styled-components": "^6.1.13",
-		"text-search-engine": "^1.2.0"
-	},
-	"peerDependencies": {
-		"react": "^16.8.0 || ^17.0.0 || ^18.0.0"
-	},
-	"devDependencies": {
-		"@types/react": "^18.3.3",
-		"rollup-plugin-peer-deps-external": "^2.2.4"
-	}
-}
