--- components/react/rollup.config.js
@@ -1,21 +1,0 @@
-import commonjs from '@rollup/plugin-commonjs'
-import resolve from '@rollup/plugin-node-resolve'
-import typescript from '@rollup/plugin-typescript'
-import peerDepsExternal from 'rollup-plugin-peer-deps-external'
-
-export default {
-	input: 'src/index.ts',
-	output: [
-		{
-			file: 'dist/index.js',
-			format: 'cjs',
-			sourcemap: true,
-		},
-		{
-			file: 'dist/index.esm.js',
-			format: 'esm',
-			sourcemap: true,
-		},
-	],
-	plugins: [peerDepsExternal(), resolve(), commonjs(), typescript({ tsconfig: './tsconfig.json' })],
-}
