--- components/react/src/index.ts
@@ -1,2 +1,0 @@
-export { HighlightComponent, TextSearch } from './TextSearchHighlight'
-export * from 'text-search-engine'
