--- components/react/tsconfig.json
@@ -1,11 +1,0 @@
-{
-	"compilerOptions": {
-		"declaration": true,
-		"declarationDir": "dist",
-		"outDir": "dist",
-		"jsx": "react",
-		"esModuleInterop": true
-		// ... 其他配置 ...
-	}
-	// ... 其他配置 ...
-}
