+++ packages/text-search-engine/package.json
@@ -7,6 +7,19 @@
 	"main": "src/index.ts",
 	"module": "src/index.ts",
 	"types": "src/index.ts",
+	"exports": {
+		".": {
+			"types": "./dist/index.d.ts",
+			"import": "./dist/index.esm.js",
+			"require": "./dist/index.js",
+			"default": "./dist/index.js"
+		},
+		"./react": {
+			"types": "./dist/react/index.d.ts",
+			"import": "./dist/react/index.esm.js",
+			"default": "./dist/react/index.esm.js"
+		}
+	},
 	"publishConfig": {
 		"main": "dist/index.js",
 		"module": "dist/index.esm.js",
