+++ packages/text-search-engine/src/react/index.ts
@@ -1,1 +1,1 @@
-export { HighlightComponent, TextSearch } from './TextSearchHighlight'
+export * from './highlight'
